Move component files to folders of their parent cmpntn
