+++ src/components/MermaidDash/MermaidDash.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState, useCallback } from 'react'
 import Header from '../Header/Header'
 import LeafletMap from '../LeafletMap'
 import FilterPane from '../FilterPane/FilterPane'
-import LoadingIndicator from '../LoadingIndicator'
-import TableView from '../TableView'
+import LoadingIndicator from './LoadingIndicator'
+import TableView from '../TableView/TableView'
 import { useLocation, useNavigate } from 'react-router-dom'
 import MetricsPane from '../MetricsPane/MetricsPane'
 import { Modal } from '../generic'
